+++ axelor-supplychain/src/main/java/com/axelor/apps/supplychain/web/SaleOrderController.java
@@ -175,7 +175,7 @@ public class SaleOrderController{
 								.param("popup-save", "false")
 								.param("forceEdit", "true")
 								.context("_showRecord", String.valueOf(saleOrder.getId()))
-								.context("supplierPartnerId", ((supplierPartner != null) ? String.valueOf(supplierPartner.getId()) : "NULL"))
+								.context("supplierPartnerId", ((supplierPartner != null) ? supplierPartner.getId() : 0L))
 								.context("saleOrderLineIdSelected", Joiner.on(",").join(saleOrderLineIdSelected))
 								.map());
 					}